Add a new attribute to the file definition so that the output format can be controlled.
Currently the supported formats are "netcdf4" and "netcdf4_classic". The "format" attribute is optional. The "netcdf4" format will be used when no format is explicitly defined. Since "netcdf4" is the format which was previously used by XIOS, existing configuration files will not be affected by this change.

If "netcdf4_classic" is used, the output file(s) will be created using the classic NetCDF format. This format can be used with the attribute "type" set to "one_file" if the NetCDF4 library was compiled with Parallel NetCDF support (--enable-pnetcdf).

git-svn-id: https://forge.ipsl.jussieu.fr/ioserver/svn/XIOS/trunk@517 d613d254-d18a-4304-b296-5125d4c4d670