Do not use the "min_buffer_size" parameter also as maximum event size.
Using the "min_buffer_size" as maximum event size caused the maximum number of bufferable events to be completely underestimated which could affect performance.

The new solution is not perfect either as it overestimates the maximum number of bufferable events in most cases. In theory this could cause deadlocks but in practice it should be mostly safe.

git-svn-id: https://forge.ipsl.jussieu.fr/ioserver/svn/XIOS/trunk@1191 d613d254-d18a-4304-b296-5125d4c4d670