Bugfix on reading curvilinear and rectilinear domains.
For both domain types a user can define either 1D or 2D longitude, latitude, and bounds. If either 1D or 2D values are given by the user, then values read from a file will be ignored. Otherwise lon, lat, and bounds will be treated as 1D for rectilinear domains and as 2D for curvilinear domains. 


git-svn-id: https://forge.ipsl.jussieu.fr/ioserver/svn/XIOS/trunk@1565 d613d254-d18a-4304-b296-5125d4c4d670

Former-commit-id: e6f07080717770858ad5d3a63c51e11d94f0334b