Interpolation enhancement : Domain area can be used to improved global conservation when there is a discrepency between model area and xios computed area.
New domain attribute : 

radius : radius af the spherical domain, used to compute area ond the sphere with a normalized radius of 1 (for remapper).

New domain_interpolate attribute :
use_area : remapper will take model computed area in order to perform a global conservation for flux integrated over the cell (mass for example).
In this case the domain attributes "area" and "radius" must be defined on the source or target domain to be taking into account.

YM 

git-svn-id: https://forge.ipsl.jussieu.fr/ioserver/svn/XIOS/trunk@1615 d613d254-d18a-4304-b296-5125d4c4d670

Former-commit-id: 7c3ff9687122f26cb2ccb080be5b0ee64d05198c